Fix exceptions

+++ OsmAnd/src/net/osmand/plus/activities/SavingTrackHelper.java
@@ -161,7 +161,7 @@ public class SavingTrackHelper extends SQLiteOpenHelper {
 		}
 
 		SQLiteDatabase db = getWritableDatabase();
-		if (db != null && warnings.isEmpty()) {
+		if (db != null && warnings.isEmpty() && db.isOpen()) {
 			try {
 				// remove all from db
 				db.execSQL("DELETE FROM " + TRACK_NAME + " WHERE " + TRACK_COL_DATE + " <= ?", new Object[] { System.currentTimeMillis() }); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
@@ -180,9 +180,13 @@ public class SavingTrackHelper extends SQLiteOpenHelper {
 	public Map<String, GPXFile> collectRecordedData() {
 		Map<String, GPXFile> data = new LinkedHashMap<String, GPXFile>();
 		SQLiteDatabase db = getReadableDatabase();
-		if(db != null) {
-			collectDBPoints(db, data);
-			collectDBTracks(db, data);
+		if (db != null && db.isOpen()) {
+			try {
+				collectDBPoints(db, data);
+				collectDBTracks(db, data);
+			} finally {
+				db.close();
+			}
 		}
 		return data;
 	}
